+++ load.h
@@ -1,5 +1,5 @@
 #include "tracer.h"
-void convertcol(long col, struct Polygon *poly);
+void convertcol(int col, struct Polygon *poly);
 
 /* Load GEO object and store in the previously allocated buffers.  Vertex */
 /* arrays are allocated during the load since the number of vertices per */
